+++ Services/BalanceReadWrite.cs
@@ -6,20 +6,20 @@ namespace AutoRepairShop.Services
 {
     class BalanceReadWrite
     {
-        public static int Balance;
+        public static double Balance;
 
-        public static int Read()
+        public static double Read()
         {
             if (new FileInfo("Balance.txt").Length == 0)
             {
                 return Balance = 1000;
             }
             string lastLine = File.ReadLines("Balance.txt").Last();
-            Int32.TryParse(lastLine, out Balance);
+            Double.TryParse(lastLine, out Balance);
             return Balance;
         }
 
-        public static void Write(int balance)
+        public static void Write(double balance)
         {
             File.WriteAllText("Balance.txt", String.Empty);
             using (StreamWriter w = File.AppendText("Balance.txt"))
